refactor(components/atoms) : common 폴더 구조 index.ts 추가
