+++ src/common/Interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './IComponent';
+
