+++ src/components/atoms/CheckBox/index.tsx
@@ -1,5 +1,5 @@
 import * as S from './style';
-import { IComponent } from 'common/Interfaces/IComponent';
+import { IComponent } from 'common';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCheckSquare } from '@fortawesome/free-solid-svg-icons';
 import { faSquare } from '@fortawesome/free-regular-svg-icons';
