+++ src/components/atoms/Input/style.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { IInputProps } from '.';
-import { ComponentMixin } from 'common/Mixins';
+import { ComponentMixin } from 'common';
 
 export const WrapInput = styled.input<IInputProps>`
   ${ComponentMixin};
