+++ src/components/atoms/TextArea/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as S from './style';
-import { IComponent } from 'common/Interfaces/IComponent';
+import { IComponent } from 'common';
 
 export interface ITextAreaProps extends IComponent {
   ref?: React.RefObject<HTMLTextAreaElement>;
